task 1

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -11,7 +11,7 @@ void print_binary(unsigned long int n)
 		_putchar('0');
 	else
 	{
-		for (i = sizeof(unsigned int) * 8 - 1; i >= 0; i--)
+		for (i = sizeof(unsigned long int) * 8 - 1; i >= 0; i--)
 		{
 			bin = (n >> i) & 1;
 			if (bin || print)
